Move json output of Lib-BDD parsed statistics into reusable function
