+++ src/apply.cpp
@@ -118,30 +118,8 @@ run_apply(int argc, char** argv)
 
       std::cout << json::field("path") << json::value(inputs_path.at(i)) << json::comma
                 << json::endl;
+      lib_bdd::print_json(stats, std::cout);
 
-      std::cout << json::field("size") << json::value(stats.size) << json::comma << json::endl;
-      std::cout << json::field("levels") << json::value(stats.levels) << json::comma << json::endl;
-      std::cout << json::field("width") << json::value(stats.width) << json::comma << json::endl;
-
-      std::cout << json::field("terminal_edges") << json::brace_open << json::endl;
-      std::cout << json::field("false") << json::value(stats.terminals[false]) << json::comma
-                << json::endl;
-      std::cout << json::field("true") << json::value(stats.terminals[true]) << json::endl;
-      std::cout << json::brace_close << json::endl;
-
-      std::cout << json::field("parent_counts") << json::brace_open << json::endl;
-      std::cout << json::field("0")
-                << json::value(stats.parent_counts[lib_bdd::stats_t::parent_count_idx::None])
-                << json::comma << json::endl;
-      std::cout << json::field("1")
-                << json::value(stats.parent_counts[lib_bdd::stats_t::parent_count_idx::One])
-                << json::comma << json::endl;
-      std::cout << json::field("2")
-                << json::value(stats.parent_counts[lib_bdd::stats_t::parent_count_idx::Two])
-                << json::comma << json::endl;
-      std::cout << json::field("3")
-                << json::value(stats.parent_counts[lib_bdd::stats_t::parent_count_idx::More])
-                << json::endl;
       std::cout << json::brace_close << json::endl;
 
       std::cout << json::brace_close;
